BIG CHANGE: Use us.metamath.org generator results!!
This commit modifies us.metamath.org so that it uses the
website generated by itself, and NO LONGER loads results from
us2.metamath.org.

Spot-checks suggest that the new generator is working just fine.
This does *not* turn off the system us2.metamath.org, so if there's
a problem, we will be able to switch back.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>